requirements.txt
Django==3.1.1
PyPDF2==1.26.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
celery==4.4.1
mysql-connector-python==8.0.21
mysqlclient==2.0.1